+++ examples/TestApi/Models/Product.cs
@@ -5,6 +5,7 @@ namespace TestApi.Models;
 // NOTE: Can be accessed at `/product`
 [RestApi]
 [GrpcApi]
+[GraphQlApi]
 public partial class Product
 {
     private readonly ILogger<Product> _logger;