+++ src/KurzSharp.EmbeddedSource/GraphQlApiAttribute.cs
@@ -1,0 +1,9 @@
+namespace KurzSharp;
+
+/// <summary>
+/// Generates A Graphql API for given Model
+/// </summary>
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct)]
+public class GraphQlApiAttribute: Attribute
+{
+}