+++ src/KurzSharp.Templates/GraphQlApi/Query.cs
@@ -1,0 +1,5 @@
+namespace KurzSharp.Templates.GraphQlApi;
+
+public class Query
+{
+}