+++ src/KurzSharp.Templates/Models/PlaceholderModel.cs
@@ -2,6 +2,7 @@
 using System.Text.Json.Serialization;
 #endif
 // NOTE: Is important in final generated code to get `LifeCycleHooks`
+// ReSharper disable once RedundantUsingDirective
 using KurzSharp;
 
 namespace KurzSharp.Templates.Models;