+++ src/KurzSharp.Templates/Models/PlaceholderModelDto.cs
@@ -8,17 +8,8 @@ namespace KurzSharp.Templates.Models;
 #if NET8_0_OR_GREATER
 [DataContract]
 #endif
-public class PlaceholderModelDto
+public record PlaceholderModelDto
 {
     [DataMember(Order = 1)]
     public Guid PlaceholderId { get; set; }
-
-#if NET8_0_OR_GREATER
-    public PlaceholderModel CopyToModel(PlaceholderModel model)
-    {
-        model.Id = PlaceholderId;
-
-        return model;
-    }
-#endif
 }