+++ src/KurzSharp/Utils/ApiKind.cs
@@ -3,5 +3,6 @@ namespace KurzSharp.Utils;
 public enum ApiKind
 {
     Rest,
-    Grpc
+    Grpc,
+    GraphQl
 }