+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added support for `GraphQlApi` with [Hotchocolate](https://github.com/ChilliCream/graphql-platform).
+
 ## [2.1.0] - 2024-06-12
 
 ### Fixed
@@ -14,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added support for `GrpcApi` with Code First `protobuf-net.Grpc`.
+- Added support for `GrpcApi` with Code First [protobuf-net.Grpc](https://github.com/protobuf-net/protobuf-net.Grpc).
 
 ## [2.0.0] - 2024-04-21